create base api class
